Add tentative definitions, setjmp, and wide characters
Three gaps that were diagnosed rather than compiled, and one more the
first of them uncovered.

A file-scope declaration with no initializer and no `extern` is a
tentative definition (C99 6.9.2). It may be written as often as you like,
and only a second initializer is a redefinition. Sema treated the first
one as final, so `int x; int x = 3;` was rejected. The symbol now records
whether a definition ever carried an initializer, and the merged type is
whichever one is complete rather than whichever came last.

setjmp and longjmp need machine state saved and restored, and kernel32
already does exactly that. RtlCaptureContext fills a CONTEXT with its
CALLER's registers and resume point, so setjmp is a macro (C99 7.13
allows it): the capture runs in the function that wrote setjmp and
records that function's own frame, still live when longjmp arrives from
deeper down. RtlRestoreContext resumes there. No assembly, and nothing
in the runtime has to know which registers the ABI calls callee-saved.

That exposed a third thing. A longjmp restores every register and touches
no memory, so C99 7.13.2.1p3 promises only volatile automatics keep their
values, and the qualifier was parsed and thrown away: a volatile local
was a value handle the backend could keep in a register. It gets storage
now, and naming it is a load.

Wide characters: `L'x'` and `L"..."` lex as literals rather than as an
identifier that happens to be one letter long. wchar_t is sixteen bits,
as on every Windows toolchain, so a wide literal is decoded from the
source's UTF-8 and re-encoded as UTF-16, and its array length counts
units rather than bytes. rt/rt_wstring.c carries wcslen and the rest.

An enumerator now goes through the real expression parser and constant
folder instead of a parallel integer-only one that knew nothing of
sizeof, which is what stopped `A = sizeof(int) + FLAG` and with it every
unity-based test.

c-testsuite goes 199 -> 204 at -O1 (00095, 00096, 00098, 00136, 00220),
and SQLite's amalgamation now compiles clean: 255,680 lines, zero errors,
and a program that links it runs and agrees with gcc. Mettle's suite
stays 647/647 built by this compiler, difftest is 51/51, and cJSON and
tinyexpr still match gcc at both optimization levels.

+++ include/stddef.h
@@ -4,7 +4,8 @@
 
 typedef long long ptrdiff_t;
 typedef unsigned long long size_t;
-typedef int wchar_t;
+/* 16 bits, as on every Windows toolchain: a wide literal is UTF-16. */
+typedef unsigned short wchar_t;
 
 #define NULL ((void *)0)
